07-项目实战/agent-memory-enhanced/requirements.txt
# 核心依赖
langgraph>=0.2.0

# Embedding 和向量检索（长期记忆需要）
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# LLM API（使用 urllib，无需 openai SDK）
